# packages/shared-types/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "seo-content-agent",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/projects/{project_id}/clustering-runs": {
      "post": {
        "tags": [
          "keyword-clusters"
        ],
        "summary": "Run Clustering",
        "operationId": "run_clustering_api_v1_projects__project_id__clustering_runs_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClusteringRunRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ClusteringRunResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/clusters": {
      "get": {
        "tags": [
          "keyword-clusters"
        ],
        "summary": "List Clusters",
        "operationId": "list_clusters_api_v1_projects__project_id__clusters_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_KeywordClusterList_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/clusters/{cluster_id}": {
      "get": {
        "tags": [
          "keyword-clusters"
        ],
        "summary": "Get Cluster",
        "operationId": "get_cluster_api_v1_projects__project_id__clusters__cluster_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "cluster_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Cluster Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_KeywordClusterDetail_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "keyword-clusters"
        ],
        "summary": "Update Cluster",
        "operationId": "update_cluster_api_v1_projects__project_id__clusters__cluster_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "cluster_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Cluster Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClusterUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_KeywordClusterDetail_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/clusters/merge": {
      "post": {
        "tags": [
          "keyword-clusters"
        ],
        "summary": "Merge Clusters",
        "operationId": "merge_clusters_api_v1_projects__project_id__clusters_merge_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClusterMergeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_KeywordClusterDetail_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/clusters/move-keywords": {
      "post": {
        "tags": [
          "keyword-clusters"
        ],
        "summary": "Move Keywords",
        "operationId": "move_keywords_api_v1_projects__project_id__clusters_move_keywords_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveKeywordsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_KeywordClusterDetail_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiResponse_ClusteringRunResponse_": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ClusteringRunResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ErrorItem"
            },
            "type": "array",
            "title": "Errors"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "ApiResponse[ClusteringRunResponse]"
      },
...
      "ApiResponse_KeywordClusterDetail_": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/KeywordClusterDetail"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ErrorItem"
            },
            "type": "array",
            "title": "Errors"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "ApiResponse[KeywordClusterDetail]"
      },
      "ApiResponse_KeywordClusterList_": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/KeywordClusterList"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ErrorItem"
            },
            "type": "array",
            "title": "Errors"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "ApiResponse[KeywordClusterList]"
      },
...
      "ClusterMergeRequest": {
        "properties": {
          "source_cluster_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "minItems": 2,
            "title": "Source Cluster Ids"
          },
          "target_cluster_name": {
            "type": "string",
            "minLength": 1,
            "title": "Target Cluster Name"
          },
          "primary_keyword_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary Keyword Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "source_cluster_ids",
          "target_cluster_name"
        ],
        "title": "ClusterMergeRequest"
      },
      "ClusterStatus": {
        "type": "string",
        "enum": [
          "proposed",
          "reviewed",
          "approved",
          "rejected"
        ],
        "title": "ClusterStatus"
      },
      "ClusterUpdateRequest": {
        "properties": {
          "cluster_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cluster Name"
          },
          "primary_keyword_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary Keyword Id"
          },
          "topic_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic Id"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ClusterStatus"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "rationale": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rationale"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ClusterUpdateRequest"
      },
      "ClusteringRunRequest": {
        "properties": {
          "similarity_threshold": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.1,
            "title": "Similarity Threshold",
            "default": 0.5
          },
          "include_archived": {
            "type": "boolean",
            "title": "Include Archived",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ClusteringRunRequest"
      },
      "ClusteringRunResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "organization_id": {
            "type": "string",
            "format": "uuid",
            "title": "Organization Id"
          },
          "project_id": {
            "type": "string",
            "format": "uuid",
            "title": "Project Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "algorithm_version": {
            "type": "string",
            "title": "Algorithm Version"
          },
          "parameters": {
            "type": "object",
            "title": "Parameters"
          },
          "keyword_count": {
            "type": "integer",
            "title": "Keyword Count"
          },
          "cluster_count": {
            "type": "integer",
            "title": "Cluster Count"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "organization_id",
          "project_id",
          "status",
          "algorithm_version",
          "parameters",
          "keyword_count",
          "cluster_count",
          "created_at"
        ],
        "title": "ClusteringRunResponse"
      },
...
      "ErrorItem": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field"
          },
          "details": {
            "type": "object",
            "title": "Details"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorItem",
        "description": "Stable machine-readable error entry."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "KeywordClusterDetail": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "organization_id": {
            "type": "string",
            "format": "uuid",
            "title": "Organization Id"
          },
          "project_id": {
            "type": "string",
            "format": "uuid",
            "title": "Project Id"
          },
          "topic_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic Id"
          },
          "topic_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic Name"
          },
          "clustering_run_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Clustering Run Id"
          },
          "cluster_name": {
            "type": "string",
            "title": "Cluster Name"
          },
          "primary_keyword_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary Keyword Id"
          },
          "primary_keyword": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary Keyword"
          },
          "intent": {
            "type": "string",
            "title": "Intent"
          },
          "cluster_score": {
            "type": "number",
            "title": "Cluster Score"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "rationale": {
            "type": "string",
            "title": "Rationale"
          },
          "member_count": {
            "type": "integer",
            "title": "Member Count",
            "default": 0
          },
          "members": {
            "items": {
              "$ref": "#/components/schemas/KeywordClusterMemberDetail"
            },
            "type": "array",
            "title": "Members"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "organization_id",
          "project_id",
          "cluster_name",
          "intent",
          "cluster_score",
          "status",
          "rationale",
          "created_at",
          "updated_at"
        ],
        "title": "KeywordClusterDetail"
      },
      "KeywordClusterList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/KeywordClusterDetail"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "KeywordClusterList"
      },
      "KeywordClusterMemberDetail": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "keyword_id": {
            "type": "string",
            "format": "uuid",
            "title": "Keyword Id"
          },
          "keyword": {
            "type": "string",
            "title": "Keyword"
          },
          "search_volume": {
            "type": "integer",
            "title": "Search Volume"
          },
          "keyword_difficulty": {
            "type": "number",
            "title": "Keyword Difficulty"
          },
          "cpc": {
            "type": "number",
            "title": "Cpc"
          },
          "intent": {
            "type": "string",
            "title": "Intent"
          },
          "priority_score": {
            "type": "number",
            "title": "Priority Score"
          },
          "business_value_score": {
            "type": "number",
            "title": "Business Value Score"
          },
          "is_primary": {
            "type": "boolean",
            "title": "Is Primary"
          },
          "similarity_score": {
            "type": "number",
            "title": "Similarity Score"
          }
        },
        "type": "object",
        "required": [
          "id",
          "keyword_id",
          "keyword",
          "search_volume",
          "keyword_difficulty",
          "cpc",
          "intent",
          "priority_score",
          "business_value_score",
          "is_primary",
          "similarity_score"
        ],
        "title": "KeywordClusterMemberDetail"
      },
...
      "MoveKeywordsRequest": {
        "properties": {
          "keyword_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "minItems": 1,
            "title": "Keyword Ids"
          },
          "target_cluster_id": {
            "type": "string",
            "format": "uuid",
            "title": "Target Cluster Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "keyword_ids",
          "target_cluster_id"
        ],
        "title": "MoveKeywordsRequest"
      },
...
      "ResponseMeta": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "has_more": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Has More"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "request_id"
        ],
        "title": "ResponseMeta",
        "description": "Request metadata shared by success and failure responses."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}